Rename file and add back learn react link
