+++ template/src/features/counter/Counter.js
